Switch from custom "random.py" to "_bootlocale.py"
It looks like we also need to customize things in "datetime" so pulling
in a custom version of every module will be trouble.  Instead we will
monkey patch everything we need in this module, which just so happens
to be the last loaded during a Python bootup (according to "strace").