Now that book is complete, remove check.py script
The script's purpose was to verify that all scripts for the book had the
same comment at the top that both named the book and correctly linked to
the online copy of the script.  Now that the book is out, the script is
no longer needed and can be moved out of the way of readers perusing
this repository.