Throw Python 2 dependencies into playground, too
And, add pycrypto!  I have been waiting for it to compile over and over
again, because I had never realized that its package name under Ubuntu
is "python-crypto" and thus does not show up when I run "apt-cache
search pycrypto".  But I found it, thanks to "apt-file search pycrypto"
and so time should now be saved for both Python 2 and Python 3.