Fix NameError “'platform_system' is not defined”
This error was stopping folks who tried to hand-build a playground,
because more modern versions of some packages were not compatible with
the “pip” from 14.04.  These version pins, taken from the installed
versions of each package in the official playground image, make sure
that home-built playgrounds (a) don’t run into the install error and
(b) remain as closely compatible as possible with the official image.